# openAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freya Salon Backend API",
    "description": "\n    ## Freya Beauty Salon Management System API\n    \n    Bu API Freya go'zallik saloni boshqaruv tizimi uchun yaratilgan.\n    \n    ### Asosiy funksiyalar:\n    - 🔐 **Autentifikatsiya va avtorizatsiya** (JWT token)\n    - 👥 **Foydalanuvchilar boshqaruvi** (Admin, User, Employee)\n    - 🏢 **Salon boshqaruvi** (Salon ma'lumotlari, xizmatlar)\n    - 📅 **Vaqt jadvali boshqaruvi** (Booking, Schedule)\n    - 💳 **To'lov tizimi** (Click.uz integratsiyasi)\n    - 📱 **SMS xabarnomalar** (Eskiz.uz integratsiyasi)\n    - 🌐 **Tarjima xizmati** (DeepL API)\n    \n    ### Texnologiyalar:\n    - **FastAPI** - Web framework\n    - **SQLAlchemy** - ORM\n    - **PostgreSQL/SQLite** - Database\n    - **JWT** - Authentication\n    - **Pydantic** - Data validation\n    \n    ### Versiya: 2.0.0\n    ",
    "contact": { "name": "Freya Development Team", "email": "dev@freya.uz" },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "servers": [
    { "url": "http://localhost:8000", "description": "Development server" },
    {
      "url": "https://freya-salon-backend.herokuapp.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/admin/salon/top": {
      "post": {
        "tags": ["Admin"],
        "summary": "Set Salon Top",
        "description": "Salonni top qilish yoki top'dan chiqarish",
        "operationId": "set_salon_top_api_admin_salon_top_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SalonTopRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SalonTopResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/admin/salons/top": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get Top Salons",
        "description": "Top salonlar ro'yxatini olish",
        "operationId": "get_top_salons_api_admin_salons_top_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/app__schemas__admin__SalonListResponse"
                  },
                  "type": "array",
                  "title": "Response Get Top Salons Api Admin Salons Top Get"
                }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/admin/salon/{salon_id}/top-history": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get Salon Top History",
        "description": "Salon top tarixini olish",
        "operationId": "get_salon_top_history_api_admin_salon__salon_id__top_history_get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "salon_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Salon Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalonTopHistoryResponse"
                  },
                  "title": "Response Get Salon Top History Api Admin Salon  Salon Id  Top History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/admin/salons": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get All Salons",
        "description": "Barcha salonlar ro'yxatini olish (pagination bilan)",
        "operationId": "get_all_salons_api_admin_salons_get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          },
          {
            "name": "is_top",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "title": "Is Top"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/app__schemas__admin__SalonListResponse"
                  },
                  "title": "Response Get All Salons Api Admin Salons Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/admin/my-salon": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get My Salon",
        "description": "Admin'ning o'z salonini olish",
        "operationId": "get_my_salon_api_admin_my_salon_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SalonDetailResponse" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/admin/send-sms": {
      "post": {
        "tags": ["Admin"],
        "summary": "Send Sms",
        "description": "SMS yuborish (admin uchun)",
        "operationId": "send_sms_api_admin_send_sms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SendSMSRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SMSResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/admin/verify-sms": {
      "post": {
        "tags": ["Admin"],
        "summary": "Verify Sms",
        "description": "SMS kodni tasdiqlash (admin uchun)",
        "operationId": "verify_sms_api_admin_verify_sms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifySMSRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SMSResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
...
    "/api/salons/": {
      "post": {
        "tags": ["Salons"],
        "summary": "Create Salon",
        "description": "Yangi salon yaratish",
        "operationId": "create_salon_api_salons__post",
        "security": [{ "HTTPBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SalonCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StandardResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Salons"],
        "summary": "Get All Salons",
        "description": "Barcha salonlarni olish",
        "operationId": "get_all_salons_api_salons__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          },
          {
            "name": "is_private",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "default": "",
              "title": "Is Private"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__salon__SalonListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SMSResponse": {
        "properties": {
          "success": { "type": "boolean", "title": "Success" },
          "message": { "type": "string", "title": "Message" },
          "data": {
            "anyOf": [{ "type": "object" }, { "type": "null" }],
            "title": "Data"
          }
        },
        "type": "object",
        "required": ["success", "message"],
        "title": "SMSResponse",
        "description": "SMS response schema"
      },
...
      "SalonTopHistoryResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "salon_id": { "type": "string", "title": "Salon Id" },
          "admin_id": { "type": "string", "title": "Admin Id" },
          "action": { "type": "string", "title": "Action" },
          "start_date": { "type": "string", "title": "Start Date" },
          "end_date": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "End Date"
          },
          "is_active": { "type": "boolean", "title": "Is Active" },
          "created_at": { "type": "string", "title": "Created At" }
        },
        "type": "object",
        "required": [
          "id",
          "salon_id",
          "admin_id",
          "action",
          "start_date",
          "is_active",
          "created_at"
        ],
        "title": "SalonTopHistoryResponse",
        "description": "Salon top history response schema"
      },
...
      "SalonTopRequest": {
        "properties": {
          "salonId": { "type": "string", "title": "Salonid" },
          "isTop": { "type": "boolean", "title": "Istop" },
          "duration": { "type": "integer", "title": "Duration", "default": 7 }
        },
        "type": "object",
        "required": ["salonId", "isTop"],
        "title": "SalonTopRequest",
        "description": "Salon top qilish request schema"
      },
      "SalonTopResponse": {
        "properties": {
          "success": { "type": "boolean", "title": "Success" },
          "message": { "type": "string", "title": "Message" },
          "data": {
            "anyOf": [{ "type": "object" }, { "type": "null" }],
            "title": "Data"
          }
        },
        "type": "object",
        "required": ["success", "message"],
        "title": "SalonTopResponse",
        "description": "Salon top qilish response schema"
      },
...
      "SendSMSRequest": {
        "properties": {
          "phone": { "type": "string", "title": "Phone" },
          "message": { "type": "string", "title": "Message" }
        },
        "type": "object",
        "required": ["phone", "message"],
        "title": "SendSMSRequest",
        "description": "SMS yuborish request schema"
      },
...
      "VerifySMSRequest": {
        "properties": {
          "phone": { "type": "string", "title": "Phone" },
          "code": { "type": "string", "title": "Code" }
        },
        "type": "object",
        "required": ["phone", "code"],
        "title": "VerifySMSRequest",
        "description": "SMS tasdiqlash request schema"
      },
...
      "app__schemas__admin__SalonListResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "address": { "type": "string", "title": "Address" },
          "phone": { "type": "string", "title": "Phone" },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email"
          },
          "is_active": { "type": "boolean", "title": "Is Active" },
          "is_top": { "type": "boolean", "title": "Is Top" },
          "rating": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Rating"
          },
          "created_at": { "type": "string", "title": "Created At" },
          "updated_at": { "type": "string", "title": "Updated At" }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "address",
          "phone",
          "is_active",
          "is_top",
          "created_at",
          "updated_at"
        ],
        "title": "SalonListResponse",
        "description": "Salon list response schema"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}